requirements.txt
# Core FastAPI and web framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Data validation and models
pydantic==2.9.2

# LangChain ecosystem for document processing and AI workflows
langchain==0.3.7
langchain-google-genai==2.0.4
langchain-community==0.3.5
langchain-openai==0.2.9
langgraph==0.2.34

# Document processing
pypdf==5.0.1
python-docx==1.1.2

# Configuration and environment
python-dotenv==1.0.1

# Logging
loguru==0.7.2

# Vector database
faiss-cpu==1.8.0

# OpenAI API
openai>=1.54.0,<2.0.0